# Conflicts:
#	rider/rider-server/src/main/scala/edp/rider/rest/util/FlowUtils.scala

+++ rider/rider-server/src/main/scala/edp/rider/rest/persistence/dal/FeedbackFlowErrDal.scala
@@ -60,6 +60,6 @@ class FeedbackFlowErrDal(feedbackFlowErrTable: TableQuery[FeedbackFlowErrTable],
   def deleteHistory(pastNdays: String) = {
     val deleteSeq = Await.result(db.run(feedbackFlowErrTable.withFilter(_.feedbackTime <= pastNdays)
       .map(_.id).result).mapTo[Seq[Long]], minTimeOut)
-    Await.result(super.deleteByFilter(_.id <= deleteSeq.max), minTimeOut)
+    if(!deleteSeq.isEmpty)Await.result(super.deleteByFilter(_.id <= deleteSeq.max), minTimeOut)
   }
 }
